Detailed Results Overview hyperparameter optim
The zip file contains .xlsx files with detailed information on the hyperparam optim. There is a .xlsx file for each phenotype which has different sheets. One sheets gives an overview on the results. Further, there is one sheet for each prediction model with the results per outerfold. Beyond that, there is a sheet with the runtime overview and tried hyperparameter combination for each outerfold and prediction model.